// src/components/Typography/types.ts
export type TypographyVariant =
	| 'h1'
	| 'h2'
	| 'h3'
	| 'h4'
	| 'h5'
	| 'h6'
	| 'subtitle1'
	| 'subtitle2'
	| 'body1'
	| 'body2'
	| 'caption'
	| 'overline';

export type TypographyAlign = 'left' | 'center' | 'right' | 'justify';

export type TypographyColor =
	| 'primary'
	| 'disabled'
	| 'info'
	| 'secondary'
	| 'success'
	| 'error'
	| 'muted'
	| 'warning';